requirements.txt
flask
python-dotenv
flask-wtf
Pillow